Fixup CDLL call to search for appropriate so file
Starting with Python 3.5 "EXT_SUFFIX" variable contains platform
information, for example ".cp35-win_amd64" or
".cpython-39-x86_64-linux-gnu.so". This is used when installing the
package with pip3 and we should probably search for that instead, as
that is what the so file is named like when installed.

Without this and installing with pip3 as follows:

   pip3 install . --user

Will at runtime throw an exception about:

   OSError: /home/nickoe/.local/lib/python3.9/site-packages/bbfec.so:
   cannot open shared object file: No such file or directory